fix(api): use full symbol with MIC in GetAsset to fix 'Mic must not be empty' error

+++ models/models.go
@@ -73,7 +73,7 @@ type AccountSummary struct {
 type SecurityInfo struct {
 	Ticker   string
 	Name     string
-	Lot      int32
+	Lot      float64
 	Currency string
 }
